refactor(ui): consolidate Tag and LabelTag into single component
Merge LabelTag color logic into Tag via variant prop system.
Tag now supports "default" (plain text) and "label" (colored pill)
variants. Migrate IssueCard consumer and delete LabelTag.

Closes #204
