+++ src/components/atoms/index.ts
@@ -1,7 +1,6 @@
 export { CI } from "./CI";
 export { Diff } from "./Diff";
 export { EmptyState } from "./EmptyState";
-export { LabelTag } from "./LabelTag";
 export { PriorityBar } from "./PriorityBar";
 export { SectionHead } from "./SectionHead";
 export { CardSkeleton, ListItemSkeleton, Skeleton, StatsBarSkeleton } from "./Skeleton";
